Trait runecoral::TensorElement

pub trait TensorElement: Sized {
    const ELEMENT_TYPE: ElementType;

    fn byte_buffer(slice: &[Self]) -> &[u8];

    fn byte_buffer_mut(slice: &mut [Self]) -> &mut [u8];
}

A Rust type that can be used as the element of a tensor.

This is an internal implementation detail and you shouldn’t need to refer to it directly.

Associated Constants

const ELEMENT_TYPE: ElementType

Required methods

fn byte_buffer(slice: &[Self]) -> &[u8]

Reinterpret a slice of this TensorElement as an immutable byte array.

fn byte_buffer_mut(slice: &mut [Self]) -> &mut [u8]

Reinterpret a slice of this TensorElement as a mutable byte array.